// front/src/enums/strategy.enum.ts
export enum Strategy {
   None = "None",
   MeanReversion = "MeanReversion",
   SimpleMovingAverageCrossover = "SimpleMovingAverageCrossover",
   ExponentialMovingAverageCrossover = "ExponentialMovingAverageCrossover",
   Momentum = "Momentum",
   LinearRegressionMachineLearning = "LinearRegressionMachineLearning",
   SupportedVectorRegressionMachineLearning = "SupportedVectorRegressionMachineLearning",
}